Major Catering Wizard Stages refactoring
