+++ src/components/pages/Catering/RecommendationsWizard.tsx
@@ -1,36 +1,21 @@
 import { ArrowRight } from '../../icons'
 import { Loading } from '../../index'
-import HighlightedMenu, { MenuContent } from '../../HighlightedMenu'
-import OptionsMenu from '../../OptionsMenu'
-import {
-  eventTypeOptions,
-  numberOfPeopleOptions,
-  selectEventType,
-  selectNumberOfPeopleIndex,
-  selectServingStyle,
-  servingStyleOptions,
-  setEventType,
-  setNumberOfPeopleIndex,
-  setServingStyle
-} from '../../../slices/recommendationsSlice'
-import RangeSlider from '../../RangeSlider'
-import AllergenOptions from './AllergenOptions'
-import BackForwardButtons from './BackForwardButtons'
-import React, { ChangeEventHandler, useCallback, useMemo, useState } from 'react'
-import { Link } from "react-router-dom"
-import { CateringContent, CateringMessage } from './common'
-import styled from '@emotion/styled'
-import CallUsButton from './CallUsButton'
-import ChatButton from './ChatButton'
-import GroupOf3People from '../../icons/GroupOf3People'
-import GroupOf6People from '../../icons/GroupOf6People'
+import HighlightedMenu from '../../HighlightedMenu'
+import React, { useCallback, useState } from 'react'
 import { useHistory } from 'react-router-dom'
-import { useDispatch, useSelector } from 'react-redux'
+import { CateringContent, CateringMessage, wizardStages } from './common'
+import styled from '@emotion/styled'
+import { useSelector } from 'react-redux'
 import { selectRevenueCenter } from '@open-tender/redux'
 import { useTheme } from '@emotion/react'
 import RecommendationsResult from './RecommendationsResult'
-import RadioButton from '../../inputs/RadioButton'
-import BigEventForm from './BigEventForm'
+import EventTypeStage from './wizardStages/EventTypeStage'
+import NumberOfPeopleStage from './wizardStages/NumberOfPeopleStage'
+import BigEventFormStage from './wizardStages/BigEventFormStage'
+import SentBigEventFormStage from './wizardStages/SentBigEventFormStage'
+import DietaryRestrictionsStage from './wizardStages/DietaryRestrictionsStage'
+import ServingStyleStage from './wizardStages/ServingStyleStage'
+import SelectMainsStage from './wizardStages/SelectMainsStage'
 
 const SkipRecommendations = styled.button`
   label: SkipRecommendations;
@@ -109,117 +94,11 @@ const AnimatedHighlightedMenu = styled(HighlightedMenu)`
   }
 `;
 
-const RangeSliderContainer = styled.div`
-  label: RangeSliderContainer;
-  
-  height: 22rem;
-`
-
-const RangeSliderMessageContainer = styled.div`
-  label: RangeSliderMessageContainer;
-`;
-
-const RangeSliderMessage = styled.div`
-  label: RangeSliderMessage;
-  
-  width: 100%;
-  text-align: center;
-  text-transform: uppercase;
-  color: ${(props) => props.theme.colors.pepper};
-  font-family: ${(props) => props.theme.fonts.headings.family};
-  font-size: 24px;
-  font-weight: ${(props) => props.theme.fonts.headings.weight};
-  margin: 5px 0px;
-
-  @media (max-width: ${(props) => props.theme.breakpoints.tablet}) {
-    font-size: 20px;
-  }
-`
-
-const RangeSliderSubMessage = styled.div`
-  label: RangeSliderSubMessage;
-  
-  text-align: center;
-`;
-
-const RangeSliderMessageButtons = styled.div`
-  label: RangeSliderMessageButtons;
-  
-  display: grid;
-  grid-template-columns: 48% 48%;
-  gap: 4%;
-  justify-content: center;
-  margin-top: 1rem;
-  
-  
-  > button, > a {
-    width: 100%
-  }
-`;
-
 const SkipIcon = styled.span`
   display: inline-flex;
   width: 12px;
   margin-left: 10px;
 `
-
-const numberOfPeopleMessage = (index:number) =>{
-  if(index < 1){
-    return <RangeSliderMessage>How many are we?</RangeSliderMessage>;
-  }
-  if(index < 4){
-    return <RangeSliderMessage>Zzz!</RangeSliderMessage>;
-  }
-  if(index < 6){
-    return <RangeSliderMessage>Wow, it's gonna be a party!</RangeSliderMessage>;
-  }
-  return (
-    <RangeSliderMessageContainer>
-      <RangeSliderMessage>Nice, let's get it going!</RangeSliderMessage>
-      <RangeSliderSubMessage>Contact us directly for big orders</RangeSliderSubMessage>
-      <RangeSliderMessageButtons>
-        <CallUsButton/>
-        <ChatButton/>
-      </RangeSliderMessageButtons>
-    </RangeSliderMessageContainer>
-  )
-}
-
-const EmptyImage = styled.span<{size?:string}>`
-  label: EmptyImage;
-  
-  display: block;
-  height: ${(props) => props.size ?? `30px`};
-  width: ${(props) => props.size ?? `30px`};
-`;
-
-interface NumberOfPeopleImageProps {
-  index: number,
-  size?: string,
-  color?: string
-}
-
-const NumberOfPeopleImage = ({
-  index, size, color
-}:NumberOfPeopleImageProps) => {
-  if(index < 1){
-    return <EmptyImage size={size}/>;
-  }
-  if(index < 4){
-    return <GroupOf3People size={size} color={color}/>;
-  }
-  return <GroupOf6People size={size} color={color}/>
-}
-
-export type wizardStages =
-  | "eventType"
-  | "numberOfPeople"
-  | "bigEventForm"
-  | "sentBigEventForm"
-  | "dietaryRestrictions"
-  | "servingStyle"
-  | "selectMains"
-  | "recommendationsResult";
 
 interface RecommendationsWizardProps{
   goBack: () => void
@@ -230,7 +109,6 @@ const RecommendationsWizard = ({
 }:RecommendationsWizardProps) => {
 
   const history = useHistory()
-  const dispatch = useDispatch()
   const theme = useTheme()
 
   const [stage, setStage] = useState<wizardStages>("eventType")
@@ -238,82 +116,11 @@ const RecommendationsWizard = ({
   const revenueCenter =
     useSelector(selectRevenueCenter)
 
-  const selectedEventTypes = useSelector(selectEventType)
-  const setSelectedEventTypes = useCallback((eventType) => dispatch(setEventType(eventType)), [dispatch])
-
-  const _numberOfPeopleIndex = useSelector(selectNumberOfPeopleIndex)
-  const _setNumberOfPeopleIndex = useCallback((numberOfPeopleIndex) => dispatch(setNumberOfPeopleIndex(numberOfPeopleIndex)), [dispatch])
-
-  const selectedServingStyle = useSelector(selectServingStyle)
-  const servingStyleOnChange = useCallback<ChangeEventHandler<HTMLInputElement>>(
-    (event)=> dispatch(setServingStyle(event.target.value))
-    , [dispatch])
-
   const skipRecommendationsOnCLick = useCallback(()=>{
     if(revenueCenter){
       history.push(`/menu/${revenueCenter.slug}`)
     }
   }, [history, revenueCenter])
-
-  const highlightedMenuOnBackClick = useMemo(() => {
-    switch(stage){
-      case "eventType":
-        return goBack
-      case "numberOfPeople":
-        return () => setStage("eventType")
-      case "bigEventForm":
-        return () => setStage("numberOfPeople")
-      case "sentBigEventForm":
-        return null
-      case "dietaryRestrictions":
-        return () => setStage("numberOfPeople")
-      case "servingStyle":
-        return () => setStage("dietaryRestrictions")
-      case "selectMains":
-        return () => setStage("servingStyle")
-      default:
-        return null
-    }
-  }, [stage, setStage, goBack])
-
-  const highlightedMenuOnForwardClick = useMemo(() => {
-    switch(stage){
-      case "eventType":
-        if(selectedEventTypes.length === 0) {
-          return null;
-        }
-        return () => setStage("numberOfPeople")
-      case "numberOfPeople":
-        if(_numberOfPeopleIndex < 1){
-          return null;
-        }
-        if(_numberOfPeopleIndex > 5){
-          return () => setStage("bigEventForm");
-        }
-        return () => setStage("dietaryRestrictions")
-      case "bigEventForm":
-        return () => setStage("sentBigEventForm")
-      case "sentBigEventForm":
-        return null
-      case "dietaryRestrictions":
-        return () => setStage("servingStyle")
-      case "servingStyle":
-        return () => setStage("selectMains")
-      case "selectMains":
-        return () => setStage("recommendationsResult")
-      default:
-        return null
-    }
-  }, [stage, setStage, selectedEventTypes, _numberOfPeopleIndex])
-
-  const highlightedMenuForwardText = useMemo(() => {
-    switch(stage){
-      case "bigEventForm":
-        return "Send!"
-      default:
-        return "Confirm"
-    }
-  }, [stage, setStage, selectedEventTypes, _numberOfPeopleIndex])
 
   return (
     <>
@@ -342,68 +149,26 @@ const RecommendationsWizard = ({
             </SkipRecommendations>
             <AnimatedHighlightedMenu>
               {stage === "eventType" &&
-                <MenuContent title="Type of event" subtitle="What kind of get together are we having?">
-                  <OptionsMenu
-                    options={eventTypeOptions}
-                    selectedOptions={selectedEventTypes}
-                    setSelectedOptions={setSelectedEventTypes}
-                    widthPercentagePerButton={100}
-                  />
-                </MenuContent>
+                <EventTypeStage goBack={goBack} setStage={setStage}/>
               }
               {stage === "numberOfPeople" &&
-                <MenuContent title="Number of people" subtitle="How big is your group?">
-                  <RangeSliderContainer>
-                    <RangeSlider options={numberOfPeopleOptions} index={_numberOfPeopleIndex} setIndex={_setNumberOfPeopleIndex}>
-                      <NumberOfPeopleImage index={_numberOfPeopleIndex} size="60px"/>
-                    </RangeSlider>
-                    {numberOfPeopleMessage(_numberOfPeopleIndex)}
-                  </RangeSliderContainer>
-                </MenuContent>
+                <NumberOfPeopleStage setStage={setStage}/>
               }
               {stage === "bigEventForm" &&
-                <MenuContent
-                  title="We want to get in touch with you!"
-                  subtitle="Leave us your details so we can contact you as soon as possible"
-                >
-                  <BigEventForm/>
-                </MenuContent>
+                <BigEventFormStage setStage={setStage}/>
               }
               {stage === "sentBigEventForm" &&
-                <MenuContent
-                  title="Request sent!"
-                  subtitle="We will contact you shortly!"
-                >
-                  <Link to={`/menu/${revenueCenter.slug}`}>Back to your Roti!</Link>
-                </MenuContent>
+                <SentBigEventFormStage />
               }
               {stage === "dietaryRestrictions" &&
-                <MenuContent title="Dietary restrictions" subtitle="Any ingredients we should rule out?">
-                  <AllergenOptions/>
-                </MenuContent>
+                <DietaryRestrictionsStage setStage={setStage}/>
               }
               {stage === "servingStyle" &&
-                <MenuContent title="Serving style" subtitle="Which serving format do you prefer?">
-                  {servingStyleOptions.map(servingStyle => (
-                    <RadioButton
-                      name="serving_style"
-                      value={selectedServingStyle}
-                      option={servingStyle}
-                      onChange={servingStyleOnChange}
-                      key={servingStyle.value}
-                    />
-                  ))}
-                </MenuContent>
+                <ServingStyleStage setStage={setStage}/>
               }
               {stage === "selectMains" &&
-                <MenuContent title="Select the mains" subtitle="How many of each mains do you want?">
-                </MenuContent>
+                <SelectMainsStage setStage={setStage}/>
               }
-              <BackForwardButtons
-                onBackClick={highlightedMenuOnBackClick}
-                onForwardClick={highlightedMenuOnForwardClick}
-                forwardText={highlightedMenuForwardText}
-              />
             </AnimatedHighlightedMenu>
           </CateringContent>)
         : <RecommendationsResult/>
